+++ src/main/java/com/codingrodent/InMemoryRecordStore/annotations/Padding.java
@@ -21,28 +21,19 @@
 *         OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 *         SOFTWARE.
 */
-package net.codingrodent.InMemoryRecordStore.record;
+package com.codingrodent.InMemoryRecordStore.annotations;
 
-import net.codingrodent.InMemoryRecordStore.annotations.*;
+import java.lang.annotation.*;
 
 /**
- * Test data record
+ * Annotation to identify unused space for padding
  */
-@PackRecord(recordByteAligned = true, fieldByteAligned = true)
-public class Record1 {
+@Target({ElementType.FIELD})
+@Inherited
+@Documented
+@Retention(RetentionPolicy.RUNTIME)
+public @interface Padding {
+    int length();
 
-    @PackField(order = 0, length = 24)
-    public Integer a = 0;
-
-    @PackField(order = 7, length = 16)
-    public int b = 7;
-
-    @PackField(order = 3, length = 16)
-    public int c = 3;
-
-    @PackField(order = 121, length = 1)
-    public boolean d = true;
-
-    @Padding(order = 4, length = 8)
-    public Void v1;
+    int order();
 }
