+++ src/main/java/com/codingrodent/InMemoryRecordStore/record/Writer.java
@@ -1,0 +1,69 @@
+/*
+* MIT License
+*
+*         Copyright (c) 2016
+*
+*         Permission is hereby granted, free of charge, to any person obtaining a copy
+*         of this software and associated documentation files (the "Software"), to deal
+*         in the Software without restriction, including without limitation the rights
+*         to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+*         copies of the Software, and to permit persons to whom the Software is
+*         furnished to do so, subject to the following conditions:
+*
+*         The above copyright notice and this permission notice shall be included in all
+*         copies or substantial portions of the Software.
+*
+*         THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+*         IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+*         FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+*         AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+*         LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+*         OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+*         SOFTWARE.
+*/
+package com.codingrodent.InMemoryRecordStore.record;
+
+import com.codingrodent.InMemoryRecordStore.core.IMemoryStore;
+
+/**
+ *
+ */
+public class Writer {
+    private RecordDescriptor recordDescriptor;
+    private IMemoryStore memoryStore;
+    private IMemoryStore.AlignmentMode mode;
+
+    /**
+     * Create a new record writer
+     *
+     * @param memoryStore      Data storage structure
+     * @param recordDescriptor Field type information
+     * @param mode             Alignment mode
+     */
+    public Writer(final IMemoryStore memoryStore, final RecordDescriptor recordDescriptor, final IMemoryStore.AlignmentMode mode) {
+        this.recordDescriptor = recordDescriptor;
+        this.memoryStore = memoryStore;
+        this.mode = mode;
+    }
+
+    /**
+     * Write a record at the specified location
+     *
+     * @param loc    Location
+     * @param record Record
+     */
+    public void putRecord(final int loc, final Object record) {
+        if (!record.getClass().equals(recordDescriptor.getClazz())) {
+            throw new IllegalArgumentException("Type supplied to writer is of the wrong type");
+        }
+        Class<?> c = record.getClass();
+        for (String fieldName : recordDescriptor.getFieldNames()) {
+            try {
+                System.out.println(c.getDeclaredField(fieldName).get(record));
+            } catch (Exception e) {
+                e.printStackTrace();
+            }
+        }
+    }
+
+}
